Add DfE design system header and styles
Add DfE design system header as all DfE products are required to adopt DfE branding and in order to bring in line with the app which has now adopted the header.

Include `dfe_frontend.css` in the styles folder as the DfE design system install package is throwing up errors.

Update `application.js` script to allow header navigation highlighting to work automatically with the new header - the script now looks for a match between page title and text contained with `*-header__navigation-link` page.