Change project external contacts view to cards
Currently external contacts are in a summary list with the ‘change’ link in the action column on the first row opposite role. This gives the appearance only the role can be changed.

This commit changes the contact from a summary list to a summary card. The summary card has action links positioned in the card’s header which better communites that these actions relate to the whole card and not parts of it.