Change top level navigation team label
Change from ‘team projects’ to ‘RCS projects’ / ‘Your team projects’.

We are doing this for user feedback to see if it lessens confusion. For example if this user is in the North West team, the ‘Team’ view will show all North West projects, but some of these could be handled by ‘RCS’. Therefore the ‘Team’ column may show ‘North West’ or ‘RCS’. We want to make it clearer which team the views within this section refer to.