Fix: make RCS see a Region column and RDO see a Team column
We make Team visible to an RDO because they do not need to see Region (e.g. it’ll always be North West) but it may be useful for RDOs to see the Team, as certain projects within the region may be handled by RCS. Similarly, for RCS, they do not need to see ‘Team’ as they will only ever see RCS projects in their views, but it will be useful for them to see region, as RCS work across regions.